move getAccount server/auth too.
